feat: Add achievement system, skill progression, and tongue twister AI analysis UI
- Implement real achievement system with 26 badges across 4 categories (streak, practice time, recordings, special)
- Add celebration overlay with confetti animation when achievements unlock
- Add EMA-based skill progression with daily caps and decay system
- Add Journey section to Progress screen with unified skill labels
- Redesign Achievements screen with category grouping
- Redesign tongue twister recording UI with blue gradient microphone and wave ring animations
- Add AI analysis results display after tongue twister recordings with animated metric bars
- Connect skill updates to warmup exercises (articulation, breathing, voice, quick)

+++ app/src/main/java/com/aivoicepower/data/local/database/dao/UserProgressDao.kt
@@ -25,7 +25,6 @@ interface UserProgressDao {
     @Query("UPDATE user_progress SET totalMinutes = totalMinutes + :minutes, updatedAt = :updatedAt WHERE id = :id")
     suspend fun addMinutes(id: String = "default", minutes: Int, updatedAt: Long = System.currentTimeMillis())
 
-    // Additional methods for v2 ViewModels
     @Query("SELECT * FROM user_progress WHERE id = :id LIMIT 1")
     suspend fun getUserProgressOnce(id: String = "default"): UserProgressEntity?
 
@@ -46,13 +45,13 @@ interface UserProgressDao {
     """)
     suspend fun updateSkillLevels(
         id: String = "default",
-        diction: Int,
-        tempo: Int,
-        intonation: Int,
-        volume: Int,
-        structure: Int,
-        confidence: Int,
-        fillerWords: Int,
+        diction: Float,
+        tempo: Float,
+        intonation: Float,
+        volume: Float,
+        structure: Float,
+        confidence: Float,
+        fillerWords: Float,
         updatedAt: Long = System.currentTimeMillis()
     )
 }
